% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{anova1}
\alias{anova1}
\title{One-way analysis of variance}
\usage{
anova1(y, f, xl = "Factor", yl = "Response Variable", step = 0:7,
  alp = 0.05, dig = 4)
}
\arguments{
\item{y}{Response variable data}

\item{f}{Factors (same length as y)}

\item{xl}{Label of x-axis, Default: 'Factor'}

\item{yl}{Label of y-axis, Default: 'Response Variable'}

\item{step}{Steps of the analysis of variance, Default: 0:7}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
One-way analysis of variance
}
\examples{
y = c(79,83,88,78,75, 81,89,91,84,86,82, 86,91,93,90,89, 76,81,82,79)
f = c(rep(150,5), rep(200,6), rep(250,5), rep(300,4))
anova1(y, f, xl="Temp", yl="Yield")
}
