% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{bayes.plot}
\alias{bayes.plot}
\title{Prior and Posterior Probabilities}
\usage{
bayes.plot(prior, post, group, cond, dcol, cex = 1, dig = 4)
}
\arguments{
\item{prior}{Prior probability distribution vector.}

\item{post}{Posterior probability distribution vector.}

\item{group}{Class names, Default: A, B, C, ...}

\item{cond}{Conditional event name, Default: F}

\item{dcol}{Bar chart colors, Default: transparent rainbow colors}

\item{cex}{Text size of the probability, Default: 1}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Displaying the Prior and the Posterior Probabilities
}
\examples{
prior = c(0.2, 0.4, 0.3, 0.1)
cond = c(0.04, 0.02, 0.01, 0.05)
tot = prior*cond
post = tot / sum(tot)
bayes.plot(prior, post)
}
