% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{bntest.plot}
\alias{bntest.plot}
\title{Exact Binomial Test}
\usage{
bntest.plot(x, n, p0, alp = 0.05, side = "two", dig = 4, dcol)
}
\arguments{
\item{x}{Vector of number of successes}

\item{n}{Sample size}

\item{p0}{Population proportion value under the null hypothesis}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{dcol}{Colors of the probability bars}
}
\value{
None.
}
\description{
Plot the Result of Exact Binomial Test
}
\examples{
bntest.plot(x=2:4, n=10, p0=0.1, side="up")
(x=6:4, n=20, p0=0.5, side="two")
}
