% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{bntest2.plot}
\alias{bntest2.plot}
\title{Three Types of Binomial Tests}
\usage{
bntest2.plot(x, n, p0, alp = 0.05, side = "two", dig = 4)
}
\arguments{
\item{x}{Number of successes}

\item{n}{Sample size}

\item{p0}{Population proportion value under the null hypothesis}

\item{alp}{Level of significance, Default: 0.05}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Plot the Results of Three Types of Binomial Tests
}
\examples{
bntest2.plot(x=15, n=200, p0=0.1, alp=0.05, side="low")
bntest2.plot(x=15, n=200, p0=0.1, alp=0.05, side="two")
}
