% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{cards2}
\alias{cards2}
\title{Sample Space of Sampling Cards}
\usage{
cards2(jokers = FALSE)
}
\arguments{
\item{jokers}{Include a joker? Default: FALSE}
}
\value{
Sample space in data frame
}
\description{
Create Sample Space of Sampling Cards
}
\examples{
cards2()
}
