% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{ch10.man}
\alias{ch10.man}
\title{Manual for Ch10. Functions}
\usage{
ch10.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch10. Inference on a Single Population
}
\examples{
ch10.man()
ch10.man(3:4)
}
