% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{ch11.man}
\alias{ch11.man}
\title{Manual for Ch11. Functions}
\usage{
ch11.man(fn = 0)
}
\arguments{
\item{fn}{Function number (0~8), Default: 0}
}
\value{
None.
}
\description{
Ch11. Inference on Two Populations
}
\examples{
ch11.man()
ch11.man(1)
}
