% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{ch12.man}
\alias{ch12.man}
\title{Manual for Ch12. Functions}
\usage{
ch12.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch12. Analysis of Categorical Data
}
\examples{
ch12.man()
ch12.man(1)
}
