% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch13-fn.R
\name{ch13.man}
\alias{ch13.man}
\title{Manual for Ch13. Functions}
\usage{
ch13.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch13. Analysis of Variance
}
\examples{
ch13.man()
ch13.man(1)
}
