% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{ch14.man}
\alias{ch14.man}
\title{Manual for Ch14. Functions}
\usage{
ch14.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch14. Correlation and Regression Analysis
}
\examples{
ch14.man()
ch14.man(1)
ch14.man(2)
ch14.man(3)
}
