% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{ch15.man}
\alias{ch15.man}
\title{Manual for Ch15. Functions}
\usage{
ch15.man(fn = 0)
}
\arguments{
\item{fn}{Function number (1,2,31,32,4,51,52,61,62,7,8), Default: 0}
}
\value{
None.
}
\description{
Ch15. Nonparametric Methods
}
\examples{
ch15.man()
ch15.man(2)
ch15.man(51)
}
