% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{ch2.man}
\alias{ch2.man}
\title{Chapter 2 Functions Manual}
\usage{
ch2.man(fn = 0)
}
\arguments{
\item{fn}{Number 0 to 7 for function. Defaults to 0.}
}
\value{
None
}
\description{
This function allows you to review the 7 functions in ch2.
}
\examples{
ch2.man()
ch2.man(3:5)
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics}
