% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{ch3.man}
\alias{ch3.man}
\title{Ch3. Probability}
\usage{
ch3.man(fn = 0)
}
\arguments{
\item{fn}{Function number to be refered. Default: 0}
}
\value{
None.
}
\description{
List of Ch3. Functions
}
\examples{
ch3.man()
}
