% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{ch4.man}
\alias{ch4.man}
\title{Manual for Ch4. Functions}
\usage{
ch4.man(fn = 0)
}
\arguments{
\item{fn}{Function Number (0-14). Default: 0}
}
\value{
None.
}
\description{
Ch4. Random Variables and Probability Distributions
}
\examples{
ch4.man()
ch4.man(7:8)
}
