% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{ch5.man}
\alias{ch5.man}
\title{Manual for Ch5}
\usage{
ch5.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch5. Expected Values of Random Variables
}
\examples{
ch5.man()
ch5.man(2)
}
