% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{ch6.man}
\alias{ch6.man}
\title{Manual for Ch.6}
\usage{
ch6.man(fn = 0)
}
\arguments{
\item{fn}{Function number, Default: 0}
}
\value{
None.
}
\description{
Ch.6 Probability Distributions of Discrete Random Variables
}
\examples{
ch6.man()
ch6.man(1:2)
}
