% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{ch7.man}
\alias{ch7.man}
\title{Manual for Ch.7 Functions}
\usage{
ch7.man(fn = 0)
}
\arguments{
\item{fn}{Function number (0~2), Default: 0}
}
\value{
None
}
\description{
Ch.7 Continuous Random Variables
}
\examples{
ch7.man()
ch7.man(1:2)
}
