% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{ch8.man}
\alias{ch8.man}
\title{Manual for Ch.8}
\usage{
ch8.man(fn = 0)
}
\arguments{
\item{fn}{Function number (0~11), Default: 0}
}
\value{
None.
}
\description{
Ch8. Normal and Related Distributions
}
\examples{
ch8.man()
ch8.man(1)
}
