% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{ch9.man}
\alias{ch9.man}
\title{Manual for Ch9. Functions}
\usage{
ch9.man(fn = 0)
}
\arguments{
\item{fn}{Function number (0~6), Default: 0}
}
\value{
None.
}
\description{
Ch9. Distributions of Sample Statistics
}
\examples{
ch9.man()
ch9.man(5)
}
