% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chi.gof1}
\alias{chi.gof1}
\title{Goodness of Fit Test from a Table}
\usage{
chi.gof1(x, p, alp = 0.05, dig = 2, dig2 = 4)
}
\arguments{
\item{x}{Table of frequency}

\item{p}{Probability vector}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point for the table, Default: 2}

\item{dig2}{Number of digits below the decimal point for the results, Default: 4}
}
\value{
list(stat=test stat, df=degree of freedom, crv=critical value, pv=p-value, tab=table)
}
\description{
Goodness of fit test from a table, when the probability is given.
}
\examples{
# Goodness-of-fit Test
x = c(31,26,22,18,13,10)
ans = chi.gof1(x)
}
