% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chi.gof2}
\alias{chi.gof2}
\title{Goodness of Fit Test for Poisson/Binomial}
\usage{
chi.gof2(tab, para, para2, dist = "pois", mc1, mc2, alp = 0.05,
  dig = 2, dig2 = 4)
}
\arguments{
\item{tab}{Table of frequency}

\item{para}{First parameter (Poisson: mean, Binomial: sample size)}

\item{para2}{Second parameter (Poisson: none, Binomial: probability)}

\item{dist}{Distribution name ('pois' or 'binom'), Default: 'pois'}

\item{mc1}{Sequential column numbers to merge (first set, lower side)}

\item{mc2}{Sequential column numbers to merge (second set, upper side)}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point for the table, Default: 2}

\item{dig2}{Number of digits below the decimal point for the results, Default: 4}
}
\value{
list(stat=test stat, df=degree of freedom, crv=critical value, pv=p-value, tab=table)
}
\description{
Goodness of Fit Test for Poisson/Binomial from a Frequency Table
}
\examples{
# Goodness-of-fit Test (Poisson/Binomial)
set.seed(1234)
x=rpois(100,5)
mytab=table(x)
chi.gof2(mytab, 5)
chi.gof2(mytab)
chi.gof2(mytab, 10, 0.5, dist="binom")
chi.gof2(mytab, 10, dist="binom")
chi.gof2(mytab, mc1=8:9)
}
