% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{chi.prob}
\alias{chi.prob}
\title{Cumulative Probability of the Chi-square Distribution}
\usage{
chi.prob(nu, xp, pup = 0.995, mt, dig = 4)
}
\arguments{
\item{nu}{Degree of freedom of the chi-square distribution}

\item{xp}{Vector of specific x-axis values}

\item{pup}{Upper limit of probabilities for quantiles, Default: 0.995}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Cumulative Probability of the Chi-square Distribution
}
\details{
DETAILS
}
\examples{
k= 1:10; nu= 5
chi.prob(nu, k)
}
