% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{chi.quant}
\alias{chi.quant}
\title{Quantile Plot of the Chi-square Distribution}
\usage{
chi.quant(nu, pv, pv2 = pv, pup = 0.999, mt, dig = 4)
}
\arguments{
\item{nu}{Degree of freedom of the chi-square distribution}

\item{pv}{Vector of probabilities for quantiles}

\item{pv2}{Vector of probabilities for specific quantiles, Default: pv}

\item{pup}{Upper limit of probabilities for quantiles, Default: 0.999}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Quantile Plot of the Chi-square Distribution
}
\examples{
nu = 5
pv = c(0.005, 0.01, 0.025, 0.05, 1:9/10, 0.95, 0.975, 0.99, 0.995)
pv2 = c(0.005, 0.01, 0.025, 0.05, 0.5, 0.95, 0.975, 0.99, 0.995)
chi.quant(nu, pv, pv2)
}
