% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{chitest.plot}
\alias{chitest.plot}
\title{Plot the Chi-square Test}
\usage{
chitest.plot(stat, df, prng, side = "two", mt, dig = 4, ppt = 20)
}
\arguments{
\item{stat}{Chi-square test statistic}

\item{df}{Degree of freedom}

\item{prng}{Range of x-axis}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{ppt}{Number of plotting points in critical region, Default: 20}
}
\value{
None.
}
\description{
Plot the Result of the Chi-square Test
}
\examples{
chi0 = var1.test(x=1.24, n=25, var0=0.8, side="up")
chitest.plot(stat=chi0$stat, df=chi0$df, side="up")

x = c(20.0, 21.5, 20.9, 19.8, 22.5, 20.3, 23.6, 18.0, 23.3, 17.8)
res = var1.test(x=x, var0=2, side="two")
chitest.plot(stat=res$stat, df=res$df, side="two")
}
