% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{chitest.plot2}
\alias{chitest.plot2}
\title{Plot the Chi-square Test}
\usage{
chitest.plot2(stat, df, alp = 0.05, side = "two", pup = 0.999,
  dig = 4, ppt = 20)
}
\arguments{
\item{stat}{Chi-square test statistic}

\item{df}{Degree of freedom}

\item{alp}{Level of significance, Default: 0.05}

\item{side}{Type of the alternative hypothesis, Default: 'two'}

\item{pup}{Maximum probability for the range of x-axis, Default: 0.999}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{ppt}{Number of plot points in the critical region, Default: 20}
}
\value{
None.
}
\description{
Plot the Result of Chi-square Test
}
\examples{
# Goodness-of-fit Test
x = c(31,26,22,18,13,10)
(ct = chisq.test(x))
chitest.plot2(stat=ct$stat, df=ct$para, side="up")
# Test of Homogeneity
x = c(20,16,29,21,14,  14,22,26,25,13,  18,24,32,18, 8,  8,18,33,16,25)
x = matrix(x, nrow=4, ncol=5, byrow=TRUE)
(ct = chisq.test(x))
chitest.plot2(stat=ct$stat, df=ct$para, side="up")
}
