% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{ciprob.sim}
\alias{ciprob.sim}
\title{Simulate the Confidence Interval for a Proportion}
\usage{
ciprob.sim(n, p = 0.5, alp = 0.05, N = 100, seed = 9857, dig = 4,
  plot = TRUE)
}
\arguments{
\item{n}{Sample size}

\item{p}{Population proportion value, Default: 0.5}

\item{alp}{Level of significance, Default: 0.05}

\item{seed}{Seed value for generating random numbers, Default: 9857}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Plot confidence intervals? Default: TRUE}

\item{n}{Sample size, Default: 100}
}
\value{
None.
}
\description{
Simulate the Confidence Interval for a Population Proportion
}
\examples{
ciprob.sim(n=16, p=0.6, alp=0.05, N=100)
ciprob.sim(n=16, p=0.6, alp=0.05, N=10000, plot=FALSE)
}
