% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{civar2.sim}
\alias{civar2.sim}
\title{Simulate the Confidence Interval for a the Ratio of Variances}
\usage{
civar2.sim(n1, n2, sig1, sig2, alp = 0.05, N = 100, seed = 9857,
  dig = 4, plot = TRUE)
}
\arguments{
\item{n1}{Sample size of population1}

\item{n2}{Sample size of population2}

\item{sig1}{Standard deviation of population1}

\item{sig2}{Standard deviation of population2}

\item{alp}{Level of significance, Default: 0.05}

\item{N}{Number of iterations, Default: 100}

\item{seed}{Seed value for generating random numbers, Default: 9857}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Logical value for plot, Default: TRUE}
}
\value{
None.
}
\description{
Simulate the Confidence Interval for a the Ratio of Population Variances
}
\examples{
civar2.sim(n1=25, n2=16, sig1=sqrt(8), sig2=2)
civar2.sim(n1=25, n2=16, sig1=sqrt(8), sig2=2, N=10000, plot=F)
}
