% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{clt.plot}
\alias{clt.plot}
\title{Diagnose the Central Limit Theorem}
\usage{
clt.plot(dist, para, para2, ns = c(10, 30, 50), d = rep(0.5, 3),
  N = 10000, seed = 9857, sigknow = TRUE)
}
\arguments{
\item{dist}{Name of population distribution ("exp","gamma","weibull","beta","norm", "t","chisq","f","pois","binom")}

\item{para}{Parameter for the first population}

\item{para2}{Parameter for the second population (if necessary)}

\item{ns}{Sample size, Default: c(10, 30, 50)}

\item{d}{Group width in histogram, Default: rep(0.5, 3)}

\item{N}{Number of iterations, Default: 10000}

\item{seed}{Seed value for generating random numbers, Default: 9857}

\item{sigknow}{Logical value for known population variance, Default: TRUE}
}
\value{

}
\description{
Diagnose the Central Limit Theorem
}
\examples{
clt.plot("exp", para=5, d=rep(0.4, 3))
clt.plot("bin", para=0.1, ns=nv, d=c(1, 0.6, 0.5))
}
