% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.cdf}
\alias{cont.cdf}
\title{CDF of a Continuous Random Variable}
\usage{
cont.cdf(FUN, low, up, xs, mt, dig = 4, pos = "bottomright")
}
\arguments{
\item{FUN}{Probability density function}

\item{low}{Lower limit of x-axis}

\item{up}{Upper limit of x-axis}

\item{xs}{Specific values of X for displaying the probability}

\item{mt}{Graph Title of the CDF}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{pos}{Legend location, Default: 'bottomright'}
}
\value{
None.
}
\description{
Cumulative Distribution Function of a Continuous Random Variable
}
\examples{
pdf = function(x) 2*exp(-2*x)*(x>0)
cont.cdf(pdf, low=-1, up=3, xs=c((1:5)*0.2, 2))
}
