% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.cond2}
\alias{cont.cond2}
\title{Conditional PDF Plot of Two Continuous Random Variables}
\usage{
cont.cond2(FUN, xc, yc, xs, ys, lo, up)
}
\arguments{
\item{FUN}{Continuous joint PDF function}

\item{xc}{Conditioning value of X}

\item{yc}{Conditioning value of Y}

\item{xs}{Specific value of X for displaying the density (given yc)}

\item{ys}{Specific value of Y for displaying the density (given xc)}

\item{lo}{Lower limit of the conditioned random variable}

\item{up}{Upper limit of the conditioned random variable}
}
\value{
Conditional PDF
}
\description{
Conditional Probability Distribution Plot of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) (x+y)*(x>=0 & x<=1)*(y>=0 & y<=1)
cont.cond2(pdf, yc=0.1, xs=0:2/2, lo=-0.2, up=1.2)
}
