% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.exp}
\alias{cont.exp}
\title{Expected Values of Continuous Random Variables}
\usage{
cont.exp(FUN, lo, up, mt, dig = 3, xn = "X", prt = FALSE,
  plot = FALSE, pos = "center")
}
\arguments{
\item{FUN}{Continuous probability density function}

\item{lo}{Lower limit of x-axis}

\item{up}{Upper limit of x-axis}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 3}

\item{xn}{Random variable name, Default: 'X'}

\item{prt}{Print expected values and variances? Default: FALSE}

\item{plot}{Plot probability density function? Default: FALSE}

\item{pos}{Legend location, Default: 'center'}
}
\value{
list(ev=Ex, std=Dx)
}
\description{
Expected Values of Continuous Random Variables
}
\examples{
fx = function(x) dnorm(x,10,2)
cont.exp(fx, 0, 20, prt=T, plot=T)
}
