% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.ind2}
\alias{cont.ind2}
\title{Independence of Two Continuous Random Variables}
\usage{
cont.ind2(FUN, lo1, up1, lo2, up2, n = 11, ep = 1e-06, prt = FALSE)
}
\arguments{
\item{FUN}{Continuous joint PDF}

\item{lo1}{Lower limit of X}

\item{up1}{Upper limit of X}

\item{lo2}{Lower limit of Y}

\item{up2}{Upper limit of Y}

\item{n}{Number of checking points between lower and upper limit, Default: 11}

\item{ep}{Error bound for comparing probablities, Default: 1e-06}

\item{prt}{Print detailed output? Default: FALSE}
}
\value{
Joint PDF
}
\description{
Determine Independence of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) (x+y)*(x>=0 & x<=1)*(y>=0 & y<=1)
cont.ind2(pdf, lo1=0, up1=1, lo2=0, up2=1)
}
