% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.jcdf}
\alias{cont.jcdf}
\title{Joint CDF of Two Continuous Random Variables}
\usage{
cont.jcdf(FUN, xs, ys, lo1 = -Inf, lo2 = -Inf)
}
\arguments{
\item{FUN}{Continuous joint PDF function}

\item{xs}{Specific values of X for displaying the cumulative probability}

\item{ys}{Specific values of Y for displaying the cumulative probability}

\item{lo1}{Lower limit of X, Default: -Inf}

\item{lo2}{Lower limit of Y, Default: -Inf}
}
\value{
Cumulative probability
}
\description{
Joint Cumulative Distribution Function of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) (x+y)*(x>0 & x<1)*(y>0 & y<1)
cont.jcdf(pdf, Inf, Inf)
cont.jcdf(pdf, 0.5, 0.5)
}
