% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.jcdfp}
\alias{cont.jcdfp}
\title{Joint CDF Plot of Two Continuous Random Variables}
\usage{
cont.jcdfp(FUN, lo1, up1, lo2, up2, mt)
}
\arguments{
\item{FUN}{Continuous joint PDF function}

\item{lo1}{Lower limit of x-axis}

\item{up1}{Upper limit of x-axis}

\item{lo2}{Lower limit of y-axis}

\item{up2}{Upper limit of y-axis}

\item{mt}{Title of the joint PDF plot}
}
\value{
None.
}
\description{
Joint CDF Plot of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) (x+y)*(x>0 & x<1)*(y>0 & y<1)
library(scatterplot3d)
cont.jcdfp(pdf, 0, 1, 0, 1)
}
