% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{cont.jexp}
\alias{cont.jexp}
\title{Expected Value of Two Continuous Random Variables}
\usage{
cont.jexp(FUN, lo1 = -Inf, up1 = Inf, lo2 = -Inf, up2 = Inf,
  dig = 4, prt = "exp")
}
\arguments{
\item{FUN}{Continuous joint probability density function}

\item{lo1}{Lower limit of X, Default: -Inf}

\item{up1}{Upper limit of X, Default: Inf}

\item{lo2}{Lower limit of Y, Default: -Inf}

\item{up2}{Upper limit of Y, Default: Inf}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{prt}{Option for detailed output in c("", "exp", "cov", "cor"), Default: ''}
}
\value{
list(Ex=E(X), Dx=D(X), Ey=E(Y), Dy=D(Y), Vxy=Cov(X,Y), Cxy=Corr(X,Y))
}
\description{
Joint pdf and Expected Value of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) 0.5*(x+3*y)*(x>0 & x<1)*(y>0 & y<1)
cont.jexp(pdf, prt="cor")
}
