% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.marg2}
\alias{cont.marg2}
\title{Marginal PDF Plot of Two Continuous Random Variables}
\usage{
cont.marg2(FUN, lo1, up1, lo2, up2, xs, ys)
}
\arguments{
\item{FUN}{Continuous joint PDF function}

\item{lo1}{Lower limit of x-axis}

\item{up1}{Upper limit of x-axis}

\item{lo2}{Lower limit of y-axis}

\item{up2}{Upper limit of y-axis}

\item{xs}{Specific value of X for displaying the probability density}

\item{ys}{Specific value of Y for displaying the probability density}
}
\value{
Marginal PDF
}
\description{
Marginal Probability Distribution Plot of Two Continuous Random Variables
}
\examples{
pdf = function(x, y) 2/7*(2*x+5*y)*(x>=0 & x<=1)*(y>=0 & y<=1)
cont.marg2(pdf, -0.2, 1.2, -0.2, 1.2, 0:2/2, 0:2/2)
}
