% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch7-fn.R
\name{cont.mpdf}
\alias{cont.mpdf}
\title{PDF and CDF for Continuous Random Variables}
\usage{
cont.mpdf(dist, lo, up, para, para2, ymax, mt, dcol, np = 100,
  pos1 = "topright", pos2 = "bottomright", xp1, xp2)
}
\arguments{
\item{dist}{Name of continuous probability distribution (one of the follows)
("exp", "gamma", "weibull", "beta", "norm", "t", "chisq", "f")}

\item{lo}{Lower limit of x-axis}

\item{up}{Upper limit of x-axis}

\item{para}{First parameter vector of PDF}

\item{para2}{Second parameter vector of PDF (if necessary)}

\item{ymax}{Upper limit of y-axis}

\item{mt}{Graph title}

\item{dcol}{Graph color vector (default as follows)
c("red", "blue", "orange2", "green4", "purple", "cyan2")}

\item{np}{Number of plot points, Default: 100}

\item{pos1}{Legend location of PDF, Default: 'topright'}

\item{pos2}{Legend location of CDF, Default: 'bottomright'}

\item{xp1}{Vector of specific x values for PDF (ignore legend)}

\item{xp2}{Vector of specific x values for CDF (ignore legend)}
}
\value{
None.
}
\description{
PDF and CDF for Continuous Random Variables
}
\examples{
lamb = 1:5
cont.mpdf("exp", 0, 3, para=lamb, ymax=5)

alp = c(0.5, 1, 2, 3); rate = 1
cont.mpdf("gamma", 0, 8, para=alp, para2=rate, ymax=1.2)

th = 1; alp = c(0.5, 1, 2, 3)
cont.mpdf("weibull", 0, 5, para=alp, para2=th, ymax=1.2)
}
