% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{cont.spdf}
\alias{cont.spdf}
\title{Plot PDF of Continuous Random Variables}
\usage{
cont.spdf(dist, lo, up, para, para2, ymax, xl, yl, dcol, np = 100, xp)
}
\arguments{
\item{dist}{Distribution name ("exp", "gamma", "weibull", "beta", "norm", "t", "chisq", "f")}

\item{lo}{Lower limit of x-axis}

\item{up}{Upper limit of x-axis}

\item{para}{First parameter vector of the distribution}

\item{para2}{Second parameter vector (except "exp", "t", "chisq")}

\item{ymax}{Upper limit of y-axis}

\item{xl}{Label vector of x-axis}

\item{yl}{Label vector of y-axis}

\item{dcol}{Graph color vector}

\item{np}{Number of plot points, Default: 100}

\item{xp}{Location vector for vertical lines}
}
\value{
None.
}
\description{
Plot (separately) the PDF of Continuous Random Variables
}
\examples{
mu = c(0,0,2,2)
sig=c(1,2,1,2)
cont.spdf("norm", -7, 7, mu, sig, xp=mu)
}
