% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{cont.trans}
\alias{cont.trans}
\title{Transformed PDF of a Continuous Random Variable}
\usage{
cont.trans(fx, TF, FTF, a, b, lo = 0, up = 1, plot = FALSE, ...)
}
\arguments{
\item{fx}{PDF of the original random variable}

\item{TF}{List of transform functions (1~8)}

\item{FTF}{List of transformed PDF (1~8)}

\item{a}{Lower limit of the original random variable for calculating P(a<X<b)}

\item{b}{Upper limit of the original random variable for calculating P(a<X<b)}

\item{lo}{Lower limit of the original random variable, Default: 0}

\item{up}{Upper limit of the original random variable, Default: 1}

\item{plot}{Plot the PDF? Default: FALSE}

\item{...}{Graphic parameters}
}
\value{
None.
}
\description{
Transformed PDF of a Continuous Random Variable
}
\examples{
fx = function(x) 2*x*(x>=0 & x<=1)
ty = function(x) 10*x - 4
tw = function(x) -10*x + 4
fy = function(y) fx((y+4)/10)/10
fw = function(w) fx((-w+4)/10)/10
cont.trans(fx, list(y=ty, w=tw), list(fy, fw), 0.3, 0.7, plot=TRUE, cex=1.3)
}
