% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{corr.mplot}
\alias{corr.mplot}
\title{Multiple Scatter Plots}
\usage{
corr.mplot(X, item, xl, yl, mt, step = 1:4, alp = 0.05, dig = 4)
}
\arguments{
\item{X}{List vector of bivariate data with 2 columns}

\item{item}{String vector of list names}

\item{xl}{Label of x-axis (default="group1")}

\item{yl}{Label of y-axis (default="group2")}

\item{mt}{Plot title}

\item{step}{Steps for correlation analysis, Default: 1:4}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Multiple Scatter Plots from Lists of Bivariate Data
}
\examples{
xd = list(cbind(iris[[1]], iris[[2]]), cbind(iris[[3]], iris[[4]]))
spec = c("Sepal", "Petal")
corr.mplot(X=xd, item=spec, xl="Length", yl="Width")
}
