% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{corr.mreg}
\alias{corr.mreg}
\title{Multiple Regression Analysis}
\usage{
corr.mreg(xd, y, form, xd2, form2, step = 0:4, newd, pvx, xrng, nx,
  alp = 0.05, dig = 4)
}
\arguments{
\item{xd}{Data frame of independent variables (explanatory variables)}

\item{y}{Vector of dependent variable (response variable) data}

\item{form}{Formula of regression model (ex: y ~ x1 + x2)}

\item{xd2}{Data frame of  independent variables in model2 (step=4, 5)}

\item{form2}{Formula of regression model2 (ex: y ~ x1 * x2) (step=4, 5)}

\item{step}{Steps of multiple regression analysis, Default: 0:4}

\item{newd}{Data frame of independent variables for step 6}

\item{pvx}{Designated number of independent variables for step 6(step=7)}

\item{xrng}{Range of independent variables for step 7}

\item{nx}{Designated number of independent variables for step 7}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Multiple Regression Analysis
}
\examples{
mpg = mtcars$mpg
xd = mtcars[4:6]
attach(xd)
form = mpg ~ hp + drat + wt
corr.mreg(xd, mpg, form, step=0:3)

form2 = mpg ~ hp * drat + wt
xd2 = data.frame(hp, drat, wt, hpd= hp * drat)
corr.mreg(xd, mpg, form, xd2, form2, step=4:5)
}
