% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{corr.plot}
\alias{corr.plot}
\title{Correlation Coefficients and Scatter Plots}
\usage{
corr.plot(X, Mt, item, dig = 4, prt = TRUE, pprt = FALSE,
  plot = FALSE)
}
\arguments{
\item{X}{Sample space vector of X}

\item{Mt}{Plot title}

\item{item}{Names of random variables}

\item{dig}{Number of effective digits, Default: 5}

\item{prt}{Print the result? Default: TRUE}

\item{pprt}{Print frequency tables? Default: FALSE}

\item{plot}{Plot the PDF and scatter plots? Default: FALSE}
}
\value{
None.
}
\description{
Correlation Coefficients and Scatter Plots of Discrete Random variables
}
\examples{
S = rolldie2(4)
item = c("Sum", "Max", "Min", "Range")
X = list()
X[[1]] = apply(S, 1, sum)
X[[2]] = apply(S, 1, max)
X[[3]] = apply(S, 1, min)
X[[4]] = X[[2]]-X[[3]]
Mt = paste("PDF of", item, "in 4 Dice")
corr.plot(X, Mt, item, pprt=T, plot=T)
}
