% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{corr.plot6}
\alias{corr.plot6}
\title{Six Scatter Plots}
\usage{
corr.plot6(m1 = 60, s1 = 10, m2 = 60, s2 = 10, r = 0.7,
  r2 = 0.8, n = 50)
}
\arguments{
\item{m1}{Mean 1. Default: 60}

\item{s1}{Standard deviation 1. Default: 10}

\item{m2}{Mean 2. Default: 60}

\item{s2}{Standard deviation 2. Default: 10}

\item{r}{Correlation Coefficient 1. Default: 0.7}

\item{r2}{Correlation Coefficient 2. Default: 0.8}

\item{n}{Number of samples. Default: 50}
}
\value{
None
}
\description{
Create 6 types of scatter plots.
}
\examples{
corr.plot6()
corr.plot6(r=0.6, r2=0.9, n=100)
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
