% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{corr.spear}
\alias{corr.spear}
\title{Pearson & Spearman Correlation Coefficient}
\usage{
corr.spear(x, y, r0 = 0, xl, yl, mt, step = 1:2, alp = 0.05,
  dig = 4)
}
\arguments{
\item{x}{Vector of x-data}

\item{y}{Vector of y-data}

\item{r0}{Correlation coefficient value under the null hypothesis, Default: 0}

\item{xl}{Name of x-data}

\item{yl}{Name of y-data}

\item{mt}{Title of scatter plot}

\item{step}{Steps of the analysis, Default: 1:2}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Pearson Correlation Coefficient & Spearman Correlation Coefficient
}
\examples{
x = c(10,7,0,1,5, 2,8,6,4,9, 3,0,2,4,6, 8)
y = c(75,77,91,64,79, 81,90,86,82,76, 89,93,80,87,83, 78)
corr.spear(x, y, r0=0, xl="Play", yl="Score", step=1:3)
}
