% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{cprt}
\alias{cprt}
\title{Conditional Probability}
\usage{
cprt(a, b)
}
\arguments{
\item{a}{Event to be conditioned.}

\item{b}{Event to be conditioning.}
}
\value{
None.
}
\description{
Calculating the Conditional Probability
}
\examples{
S = rolldie2(2)
A = subset(S, (X1+X2) >=4)
B = subset(S, (X1+X2) >=8)
cprt(B, A)
}
