% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_group.R
\name{des_group}
\alias{des_group}
\title{Deskriptivna statistika po grupama}
\usage{
des_group(df)
}
\arguments{
\item{df}{Data frame koji sadrži podatke.}
}
\value{
Data frame sa stupcima:
\item{VAR}{Naziv varijable}
\item{GROUP}{Naziv grupe}
\item{N}{Broj entiteta za svaku grupu}
\item{MEAN}{Aritmetička sredina za svaku grupu}
\item{SD}{Standardna devijacija za svaku grupu}
\item{SE}{Standardna greška aritmetičke sredine za svaku grupu}
}
\description{
Funkcija računa osnovne deskriptivne pokazatelje (N, MEAN, SD, SE) za numeričke
varijable grupirane po kategorijalnoj varijabli. Korisnik bira varijable putem interaktivnog izbornika.
}
\examples{
rezultat <- des_group(df)

}
