% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des_par.R
\name{des_par}
\alias{des_par}
\title{Deskriptivna statistika}
\usage{
des_par(data)
}
\arguments{
\item{data}{A \code{data.frame} koji sadrži numeričke varijable za analizu.}
}
\value{
\code{data.frame} sa sljedećim stupcima:
\itemize{
\item \strong{VARS} - naziv varijable
\item \strong{n} - broj entiteta
\item \strong{MEAN} - aritmetička sredina
\item \strong{MEDIAN} - medijan
\item \strong{SD} - standardna devijacija
\item \strong{CV} - koeficijent varijabilnosti (u \%)
\item \strong{MIN} - minimalna vrijednost
\item \strong{MAX} - maksimalna vrijednost
\item \strong{RANGE} - totalni raspon (max - min)
\item \strong{SKEW} - skewnes
\item \strong{KURT} - kurtosis
}
}
\description{
Funkcija \code{des_par} izračunava osnovne deskriptivne pokazatelje (n, mean, median, sd,
coefficient of variation, min, max, range, skewness, kurtosis) za numeričke varijable
u zadanom \code{data.frame}-u. Korisnik bira varijable putem interaktivnog izbornika.
}
\details{
Funkcija koristi paket \code{psych}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
}
\examples{
\dontrun{
# Primjer korištenja
des_par(mtcars)
}

}
