% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.cdf}
\alias{disc.cdf}
\title{CDF of a Discrete Random Variable}
\usage{
disc.cdf(xv, xp, mt, cpt = 1.2, cex = 1, dig = 3)
}
\arguments{
\item{xv}{Vector of values of the discrete random variable}

\item{xp}{Vector of the discrete PDF}

\item{mt}{Graph Title of the CDF}

\item{cpt}{Text size of the limit points, Default: 1.2}

\item{cex}{Text size of the probability, Default: 1}

\item{dig}{Number of digits below the decimal point, Default: 3}
}
\value{
None.
}
\description{
Cumulative Distribution Function of a Discrete Random Variable
}
\examples{
disc.cdf(0:3, choose(3, 0:3))
}
