% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.cond2}
\alias{disc.cond2}
\title{Conditional PDF of Discrete Random Variables}
\usage{
disc.cond2(tabXY, Xs, Ys, prt = TRUE, plot = FALSE, dig = 5,
  dig2 = 4)
}
\arguments{
\item{tabXY}{Joint frequency table of two random variables}

\item{Xs}{Conditioning value of X}

\item{Ys}{Conditioning value of Y}

\item{prt}{Print the marginal frequency and probability? Default: TRUE}

\item{plot}{Plot the marginal PDF? Default: FALSE}

\item{dig}{Number of digits below the decimal point in the console, Default: 5}

\item{dig2}{Number of digits below the decimal point in the graph, Default: 4}
}
\value{
Conditional Frequency and PDF
}
\description{
Conditional Probability Distribution of Discrete Random Variables
}
\examples{
fxy = with(mtcars, table(cyl, carb))
disc.cond2(fxy, Ys=1:4, plot=TRUE)
}
