% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{disc.exp}
\alias{disc.exp}
\title{Expected Value of a Discrete Random Variable}
\usage{
disc.exp(xv, xf, mt, dig = 3, prt = TRUE, plot = FALSE,
  pos = "topright")
}
\arguments{
\item{xv}{Vector of random variable values}

\item{xf}{Vector of probability (or frequency) distribution}

\item{mt}{graph title}

\item{dig}{Number of digits below the decimal point (default=3)}

\item{prt}{Print detailed output? Default: FALSE}

\item{plot}{Plot the PDF? Default: FALSE}

\item{pos}{Legend location, Default: 'topright'}
}
\value{
list(Ex=expected value, Dx=standard deviation, Vx=variance)
}
\description{
Expected Value of a Discrete Random Variable
}
\examples{
p = c(1, 3, 3, 1)
x = c(-3, -1, 1, 3)*100
disc.exp(x, p)
}
