% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.ind2}
\alias{disc.ind2}
\title{Independence of Two Discrete Random Variables}
\usage{
disc.ind2(X, Y, prt = TRUE)
}
\arguments{
\item{X}{Sample space vector of X}

\item{Y}{Sample space vector of Y}

\item{prt}{Print detailed output? Default: TRUE}
}
\value{
Joint PDF
}
\description{
Determine Independence of Two Discrete Random Variables
}
\examples{
S = rolldie2(4)
sum3 = function(x) sum(x>=3)
X = apply(S, 1, sum3)
even = function(x) sum(x \%\% 2 ==0)
Y = apply(S,1, even)
disc.ind2(X, Y)
}
