% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch5-fn.R
\name{disc.jexp}
\alias{disc.jexp}
\title{Expected Values of Two Discrete Random Variables}
\usage{
disc.jexp(tabXY, Xn = "X", Yn = "Y", prt = "exp", pprt = FALSE,
  dig = 4)
}
\arguments{
\item{tabXY}{Joint frequency (or probability) table}

\item{Xn}{Name of X, Default: 'X'}

\item{Yn}{Name of Y, Default: 'Y'}

\item{prt}{Option for detailed output in c("", "exp", "cov", "cor"), Default: 'exp'}

\item{pprt}{Plot the joint & marginal PDF? Default: FALSE}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
list(Ex=E(X), Dx=D(X), Ey=E(Y), Dy=D(Y), Vxy=Cov(X,Y), Cxy=Corr(X,Y))
}
\description{
Joint Probabilities and Expected Values of Two Discrete Random Variables
}
\examples{
S = rolldie2(2)
X = apply(S, 1, max)
Y = apply(S, 1, min)
tXY = table(X, Y)
disc.jexp(tXY, prt="cor", pprt=TRUE)
}
