% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{disc.joint2}
\alias{disc.joint2}
\title{Joint PDF of Two Discrete Random variable}
\usage{
disc.joint2(X, Y, prt = TRUE, plot = FALSE, dig = 4, dig2 = 3,
  ep = 0)
}
\arguments{
\item{X}{Sample space vector of the first random variable}

\item{Y}{Sample space vector of the second random variable}

\item{prt}{Print the joint frequency and probability? Default: TRUE}

\item{plot}{Plot the joint PDF? Default: FALSE}

\item{dig}{Number of digits below the decimal point in the console, Default: 4}

\item{dig2}{Number of digits below the decimal point in the graph, Default: 3}

\item{ep}{Minimum value for displaying the joint probability, Default: 0}
}
\value{
Joint PDF
}
\description{
Joint Probability Distribution of Two Discrete Random variable
}
\examples{
S = rolldie2(4)
X = apply(S, 1, max)
Y = apply(S, 1, min)
disc.joint2(X, Y)
library(scatterplot3d)
disc.joint2(X, Y, prt=FALSE, plot=TRUE)
}
