% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{disc.mexp}
\alias{disc.mexp}
\title{Expected Values of Discrete Random Variables}
\usage{
disc.mexp(xv, fx, fx2, fx3, mt, dig = 3, del = 0.2, prt = TRUE,
  plot = TRUE)
}
\arguments{
\item{xv}{Vector of random variable values}

\item{fx}{List of probability distributions (2~9)}

\item{fx2}{Second list of probability distributions (same number of fx)}

\item{fx3}{Third list of probability distributions (same number of fx)}

\item{mt}{Vector of plot titles}

\item{dig}{Number of digits below decimal point, Default: 3}

\item{del}{Distance between the probability distribution plots, Default: 0.2}

\item{prt}{Print the expected values and variances? Default: TRUE}

\item{plot}{Plot the probability distributions? Default: TRUE}
}
\value{
list(Ex, Vx, Dx)
}
\description{
Expected Values of Discrete Random Variables
}
\examples{
# Binomial distributions
n = 10;  p=c(0.2, 0.5, 0.8); x = 0:n
fx1 = fx2 = list()
for (i in 1:3) fx1[[i]] = dbinom(x, n, p[i])
mt1 = paste0("B(10,", p, ")")
disc.mexp(x, fx1, mt=mt1)
# Binomial vs. Hypergeometric
N = 50;  S =c(10, 25, 40); n = 10; x = 0:n
for (i in 1:3) fx2[[i]] = dhyper(x, S[i], N-S[i], n)
mt12 = paste0("HG(10,50,", S,"):B(10,", p,")")
disc.mexp(x, fx2, fx1, mt=mt12)
}
