% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discra_analysis.R
\name{discra_analysis}
\alias{discra_analysis}
\title{Diskriminantna analiza}
\usage{
discra_analysis(df)
}
\arguments{
\item{df}{data.frame koji sadrži podatke za analizu}
}
\value{
Lista koja sadrži:
\item{df_res}{Rezultati kanoničke diskriminantne analize}
\item{structure}{Struktura diskriminantnih varijabli}
\item{centroids}{Centroidi grupa s frekvencijama}
\item{freq_table}{Tablica frekvencija LDA klasifikacije}
\item{prop_table}{Tablica postotaka LDA klasifikacije (po redcima)}
}
\description{
Ova funkcija izvodi diskriminantnu analizu s interaktivnim izborom kvalitativne (grupne) i kvantitativnih varijabli
}
\examples{
\dontrun{
rezultati <- pokreni_diskriminant(df)
print(rezultati$df_res)
print(rezultati$structure)
print(rezultati$centroids)
print(rezultati$freq_table)
print(rezultati$prop_table)
}
}
