% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtf.R
\name{dtf}
\alias{dtf}
\title{Prikaz data frame-a u DT tablici sa rownames = FALSE}
\usage{
dtf(x)
}
\arguments{
\item{x}{Data frame koji se prikazuje.}
}
\value{
Objekt klase \code{datatable}.
}
\description{
Prikaz data frame-a u DT tablici sa rownames = FALSE
}
\examples{
dtf(mtcars)
}
