% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtt.R
\name{dtt}
\alias{dtt}
\title{Prikaz data frame-a u DT tablici sa rownames = TRUE}
\usage{
dtt(x)
}
\arguments{
\item{x}{Data frame koji se prikazuje.}
}
\value{
Objekt klase \code{datatable}.
}
\description{
Prikaz data frame-a u DT tablici sa rownames = TRUE
}
\examples{
dtt(mtcars)
}
