% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{f.prob}
\alias{f.prob}
\title{Cumulative Probability of the F-distribution}
\usage{
f.prob(nu1 = 5, nu2 = 5, xp, mt, dig = 4)
}
\arguments{
\item{nu1}{Numerator degree of freedom, Default: 5}

\item{nu2}{Denominator degree of freedom, Default: 5}

\item{xp}{Vector of x-axis values}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Cumulative Probability of the F-distribution
}
\examples{
k= 1:7
nu1 = 8; nu2= 5
f.prob(nu1, nu2, k)
}
