% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{friedman.plot}
\alias{friedman.plot}
\title{Friedman Test}
\usage{
friedman.plot(x, a, b, dig = 4, plot = FALSE)
}
\arguments{
\item{x}{Data vector}

\item{a}{Vector of factor levels}

\item{b}{Vector of block levels}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Plot test results? Default: FALSE}
}
\value{
None.
}
\description{
Friedman Test with a Plot
}
\examples{
x = c(71,77,75,79,88, 70,94,74,74,83, 72,95,77,80,90, 94,64,76,76,89)
a = rep(1:4, each=5)
b = rep(1:5, 4)
friedman.plot(x, a, b, plot=TRUE)
}
