% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{ftest.plot}
\alias{ftest.plot}
\title{Plot the PDF of the F-test Statistic}
\usage{
ftest.plot(fstat, deg, pmax = 0.995, side = "two")
}
\arguments{
\item{fstat}{F-test statistic for the ratio of two population variances}

\item{deg}{Vector of degree of freedoms}

\item{pmax}{Maximum probability for quantiles in x-axis, Default: 0.995}

\item{side}{Type of the alternative hypothesis, Default: 'two'}
}
\value{
None.
}
\description{
Plot the PDF of the F-test Statistic
}
\examples{
vo = var.test(rnorm(20, 10, 2.4), rnorm(25, 12, 1.8))
ftest.plot(fstat=vo$stat, deg=as.vector(vo$para), pmax=0.9999, side="two")
}
