% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_2char_vars.R
\name{group_2char_var}
\alias{group_2char_var}
\title{Grupiranje dviju kategorijalnih varijabli}
\usage{
group_2char_var(x, y)
}
\arguments{
\item{x, }{y Varijable se odabiru putem interaktivnog izbornika.}
}
\value{
Lista sa sljedećim komponentama:
\itemize{
\item \strong{ctm} - kontigencijska tablica s dodanim marginama (Sum row/col)
\item \strong{ctprm} - postotci po redcima s dodanim sumama
\item \strong{hit} - tekstualni rezultat chi-square testa s Yatesovom korekcijom
\item \strong{hif} - tekstualni rezultat Pearsonovog chi-square testa
\item \strong{bar_stack} - grafikon s razdjeljenim stupcima (stacked bar)
}
}
\description{
Funkcija \code{group_2char_var} omogućuje odabir dviju kategorijalnih varijabli
iz \code{data.frame}-a, izračunava kontigencijsku tablicu, postotke po redcima, te provodi
Pearsonov chi-square test i chi-square test s Yatesovom korekcijom te stacked bar grafikon pomoću \code{plotly}.
}
\details{
Funkcija koristi paket \code{plotly}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
Varijable se odabiru putem interaktivnog izbornika, a funkcija radi samo s faktorskim
ili karakter varijablama.
}
\examples{
\dontrun{
# Primjer korištenja s data.frame-om df
group_2char_var(df)
}

}
