% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_char_var.R
\name{group_char_var}
\alias{group_char_var}
\title{Grupiranje kategorijalne varijable}
\usage{
group_char_var(x)
}
\arguments{
\item{x}{Varijabla se biraju putem interaktivnog izbornika.}
}
\value{
Lista sa sljedećim komponentama:
\itemize{
\item \strong{ft} - tablica frekvencija
\item \strong{hi} - tekstualni rezultat chi-square testa
\item \strong{bar} - grafikon s jednostavnim stupcima (bar chart)
\item \strong{pie} - strukturni krug (pie chart)
}
}
\description{
Funkcija \code{group_char_var} omogućuje odabir jedne kategorijalne varijable iz
\code{data.frame}-a, izračunava frekvencije i postotke po kategorijama,
očekivane frekvencije za uniformnu distribuciju, te provodi chi-square test.
Kreira interaktivni grafikon s jednostavnim stupcima (bar chart) i strukturni krug (pie chart) pomoću \code{plotly}.
}
\details{
Funkcija koristi paket \code{plotly}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
Varijabla se biraju putem interaktivnog izbornika, a funkcija radi samo s kategorijalnim varijablama.
}
\examples{
\dontrun{
# Primjer korištenja s data.frame-om df
group_char_var(df)
}

}
