% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_num_var.R
\name{group_num_var}
\alias{group_num_var}
\title{Grupiranje numeričke varijable}
\usage{
group_num_var(x, br_razreda = NULL)
}
\arguments{
\item{x}{varijabla se odabire putem interaktivnog izbornika.}

\item{br_razreda}{Broj razreda (bins) za histogram. Ako je \code{NULL}, koristi se Sturgesovo pravilo.}
}
\value{
Lista sa sljedećim komponentama:
\itemize{
\item \strong{tf} - tablica frekvencija, postotaka i kumulativnih vrijednosti
\item \strong{hist} - histogram frekvencija
\item \strong{poli} - poligon frekvencija
}
}
\description{
Funkcija \code{group_num_var} omogućuje odabir jedne numeričke varijable iz
\code{data.frame}-a, kreira frekvencijsku tablicu prema rezredima (bins) i
prikazuje histogram te frekvencijski poligon koristeći \code{plotly}.
}
\details{
Funkcija koristi paket \code{plotly}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
Varijabla se odabire putem interaktivnog izbornika, a broj razreda se može zadati
ručno ili koristiti Sturgesovo pravilo za automatski izračun.
}
\examples{
\dontrun{
# Primjer korištenja s data.frame-om df
group_num_var(df)
group_num_var(df, br_razreda = 8)
}

}
