% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{hyp.sample}
\alias{hyp.sample}
\title{PDF of Successes from Finite Population}
\usage{
hyp.sample(npop, ndef, nsamp, cex = 0.8, dig = 4)
}
\arguments{
\item{npop}{Population size}

\item{ndef}{Number of success items in the population}

\item{nsamp}{Sample size}

\item{cex}{Text size, Default: 0.8}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
the PDF
}
\description{
Probability Distribution of the Number of Successes from Finite Population
}
\examples{
 hyp.sample(50, 8, 10)
}
