% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_data}
\alias{import_data}
\title{Interaktivno učitavanje CSV, Excel ili ODS datoteka}
\usage{
import_data()
}
\value{
Data frame s učitanim podacima.
}
\description{
Interaktivno učitavanje CSV, Excel ili ODS datoteka
}
\examples{
\dontrun{
df <- import_data()
}
}
