% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data2.R
\name{import_data2}
\alias{import_data2}
\title{Interaktivno učitavanje CSV, Excel ili ODS datoteka iz sheet = 1}
\usage{
import_data2()
}
\value{
Data frame s učitanim podacima.
}
\description{
Interaktivno učitavanje CSV, Excel ili ODS datoteka iz sheet = 1
}
\examples{
\dontrun{
df <- import_data2()
}
}
