% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{indep.event}
\alias{indep.event}
\title{Independence of Random Variables}
\usage{
indep.event(X, Y, N, ep = 1e-06)
}
\arguments{
\item{X}{First random variable.}

\item{Y}{Second random variable.}

\item{N}{Size of the sample space.}

\item{ep}{Precision limit. Default: 1e-06}
}
\value{
Probablities to be compared.
}
\description{
Determining Independence of Two Discrete Random Variables
}
\examples{
indep.event(A, B, nrow(S))
}
