% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{kruswall.plot}
\alias{kruswall.plot}
\title{Kruskal Wallis Test}
\usage{
kruswall.plot(x, y, dig = 4, plot = FALSE)
}
\arguments{
\item{x}{Data vector}

\item{y}{Vector of factor levels}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Plot test results? Default: FALSE}
}
\value{
None.
}
\description{
Kruskal Wallis Test with a Plot
}
\examples{
x = c(4.6,10.6,8.0,25.0,7.1, 2.9,10.1,3.2,3.8,6.6, 6.8,9.4,26.5,12.8,8.3, 3.4,3.9,6.0,8.6,5.0)
y = rep(1:4, each=5)
kruswall.plot(x, y, plot=TRUE)
}
