% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{mean2.ci}
\alias{mean2.ci}
\title{CI for the Difference of Two Means}
\usage{
mean2.ci(xb1, xb2, s1, s2, n1, n2, pvar = "equal", alp = 0.05,
  dig = 4)
}
\arguments{
\item{xb1}{Sample mean of population1 (or sample data)}

\item{xb2}{Sample mean of population2 (or sample data)}

\item{s1}{Standard deviation of population1 (optional for unknown variance)}

\item{s2}{Standard deviation of population2 (optional for unknown variance)}

\item{n1}{Sample size of population1 (unnecessary if data are given)}

\item{n2}{Sample size of population2 (unnecessary if data are given)}

\item{pvar}{Status of variance (one of "known", "equal", "unequal"), Default: 'equal'}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for the Difference of Two Population Means
}
\examples{
mean2.ci(xb1=198.5, xb2=201.3, s1=5, s2=5, n1=25, n2=34, pvar="known")
mean2.ci(xb1=198.5, xb2=201.3, s1=4.8, s2=5.1, n1=25, n2=34, pvar="equal")
mean2.ci(xb1=198.5, xb2=201.3, s1=2.8, s2=5.5, n1=25, n2=34, pvar="unequal")

x = rnorm(20, 199, 2)
y = rnorm(25, 200, 2)
mean2.ci(x, y, pvar="equal")
}
