% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{meanchar.plot}
\alias{meanchar.plot}
\title{Operating Characteristic Curves of the Test for a Mean}
\usage{
meanchar.plot(mu0, mu1, sig, nv, alp = 0.05, prng, side = "two", mt,
  dig = 4)
}
\arguments{
\item{mu0}{Population mean value under the null hypothesis}

\item{mu1}{Vector of population mean values for which the power should be calculated}

\item{sig}{Population standard deviation}

\item{nv}{Sample size vector}

\item{alp}{Level of significance, Default: 0.05}

\item{prng}{Range of x-axis}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Operating Characteristic Curves of the Test for a Population Mean
}
\examples{
n = c(10, 30, 50, 100)
meanchar.plot(mu0=100, mu1=98, sig=5, nv=n, side="low")
meanchar.plot(mu0=100, mu1=98, sig=5, nv=n, side="two")
}
