% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{meantest1.plot}
\alias{meantest1.plot}
\title{Hypothesis Test for a Mean (known variance)}
\usage{
meantest1.plot(xb, mu0, sig, n, prng, side = "two", mt, dig = 4,
  xlab = "Sample Mean")
}
\arguments{
\item{xb}{Sample mean, or sample data}

\item{mu0}{Population mean value under the null hypothesis}

\item{sig}{Population standard deviation}

\item{n}{Sample size (unnecessary when sample data are given)}

\item{prng}{Range of x-axis}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{xlab}{Label of x-axis, Default: 'Sample Mean'}
}
\value{
None.
}
\description{
Hypothesis Test for a Population Mean (known variance)
}
\examples{
meantest1.plot(xb=12.64, mu0=12.5, sig=0.5, n=40, side="up")
meantest1.plot(rnorm(40, 12.6, 0.5), mu0=12.5, sig=0.5, side="up")

}
