% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{meantest2.plot}
\alias{meantest2.plot}
\title{Hypothesis Test for a Mean (unknown variance)}
\usage{
meantest2.plot(xb, mu0, sig, n, prng = c(-4, 4), side = "two", mt,
  dig = 4)
}
\arguments{
\item{xb}{Sample mean, or sample data}

\item{mu0}{Population mean value under the null hypothesis}

\item{sig}{Sample standard deviation (unnecessary when sample data are given)}

\item{n}{Sample size (unnecessary when sample data are given)}

\item{prng}{Range of x-axis, Default: c(-4, 4)}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Hypothesis Test for a Population Mean (unknown variance)
}
\examples{
meantest2.plot(xb=12.65, mu0=12.5, sig=0.57, n=40, side="up")
meantest2.plot(rnorm(40, 12.65, 0.57), side="two")
}
