% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch12-fn.R
\name{mosaic2}
\alias{mosaic2}
\title{Mosaic Plot}
\usage{
mosaic2(tab, mt, dig = 4, resid = TRUE)
}
\arguments{
\item{tab}{Data table}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{resid}{Display Pearson residuals? Default: TRUE}
}
\value{
Object from chisq.test()
}
\description{
Mosaic Plot
}
\examples{
require(vcd)
x = c(39,18,12,31,14, 35,23,18,35,13,  27,16,17,24,8,  9,12,8,19,22)
x = matrix(x, nrow=4, ncol=5, byrow=TRUE)
Subject = c("Kor", "Eng", "Math", "Etc")
Hope = c("Sam", "Pub", "Exp", "Sal", "Etc")
t1 = as.table(x)
dimnames(t1) = list(Subject=Subject, Hope=Hope)
win.graph(7, 6)
mosaic2(tab=t1, mt="Students' Favorite Subject and Hope")
}
