% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_reg.R
\name{multi_reg}
\alias{multi_reg}
\title{Višestruka regresijska analiza}
\usage{
multi_reg(df)
}
\arguments{
\item{df}{Data frame koji sadrži numeričke varijable za regresijsku analizu.}
}
\value{
Lista s elementima:
\item{results_table}{Data frame koji sadrži regresijske koeficijente, SE, Beta, Part_R, R, P, Tolerance, t-vrijednosti i p-vrijednosti.}
\item{plot}{Plotly bar graf standardiziranih koeficijenata.}
\item{R_squared}{Koeficijent determinacije (R^2).}
\item{SEE}{Standardnu pogrešku prognoze.}
\item{F_value}{F statistika modela.}
\item{F_p_value}{p-vrijednost F statistike.}
\item{Durbin_Watson}{Rezultat Durbin-Watson testa za autokorelaciju reziduala.}
}
\description{
Funkcija izvodi višestruku regresijsku analizu na tablici podataka.
}
\examples{
\dontrun{
  output <- regression_analysis_manual(df)
  output$results_table
  output$plot
}
}
