% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch6-fn.R
\name{multinorm.plot}
\alias{multinorm.plot}
\title{Multinomial PDF Plot}
\usage{
multinorm.plot(ps, size)
}
\arguments{
\item{ps}{Probability (or proportion) matrix with column for each group}

\item{size}{Sample size}
}
\value{
Matrix of multinomial PDF
}
\description{
Graphic Display of Multinomial Probability Distribution Function
}
\examples{
library(scatterplot3d)
ps = matrix(c(1,1,8, 1,5,4, 4,4,2, 1,1,1), nrow=4, ncol=3, byrow=T)
multinorm.plot(ps, 5)
}
