% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{norm.diag}
\alias{norm.diag}
\title{Diagnosis of Normality}
\usage{
norm.diag(x, xrng, by = 1, dig = 4, dc = c("cyan", 2, 4))
}
\arguments{
\item{x}{Data vector}

\item{xrng}{Range of x-axis, Default: c(mean-3stdev, mean+3stdev)}

\item{by}{Histogram class interval, Default: 1}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{dc}{Color vector, Default: c("cyan", 2, 4)}
}
\value{
None.
}
\description{
Investigate the Normality of Data
}
\examples{
(x = c(1,2,5,7, rep(8,7), rep(9,5), rep(10,4)))
norm.diag(x)
}
