% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{norm.sim}
\alias{norm.sim}
\title{Simulation for the Normal Distribution}
\usage{
norm.sim(ns, mu = 0, sig = 1, N = 10000, ng = 50, seed = 9857,
  dig = 4)
}
\arguments{
\item{ns}{Sample size}

\item{mu}{Expected value, Default: 0}

\item{sig}{Standard deviation, Default: 1}

\item{N}{Number of iterations, Default: 10000}

\item{ng}{Number of classes in histogram, Default: 50}

\item{seed}{Seed value for generating random numbers, Default: 9857}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Simulation for the Normal Distribution
}
\examples{
norm.sim(ns=10, mu=100, sig=10, N=10000)
}
