% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch9-fn.R
\name{norm.spn}
\alias{norm.spn}
\title{Minimum Number of Samples}
\usage{
norm.spn(kp, alp, lo = 0.1, up = 1, mt, dcol, log = TRUE)
}
\arguments{
\item{kp}{Error limit in multiples of the standard deviation}

\item{alp}{Level of significance vector}

\item{lo}{Lower limit of the error limit, Default: 0.1}

\item{up}{Upper limit of the error limit, Default: 1}

\item{mt}{Plot title}

\item{dcol}{Line color (default=rainbow())}

\item{log}{Logical value for log-scaling y-axis, Default: TRUE}
}
\value{
None.
}
\description{
Minimum Number of Samples from Normal Population
}
\examples{
alp = c(0.01, 0.05, 0.1)
dcol = c(2, 4, "green4")
norm.spn(kp=0.4, alp, dcol=dcol)
}
