% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{norm.trans}
\alias{norm.trans}
\title{Standardization of the Normal Distribution}
\usage{
norm.trans(mu, sig, a, b, mt1, mt0, dig = 4, span = 3, np = 100)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sig}{Standard deviation of the normal distribution}

\item{a}{Lower limit of X for calculating probability P(a<X<b)}

\item{b}{Upper limit of X for calculating probability P(a<X<b)}

\item{mt1}{Title of the normal distribution probability plot}

\item{mt0}{Title of the standard normal distribution probability plot}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{span}{Range of x-axis in [mu-span*sig, mu+span*sig], Default: 3}

\item{np}{Number of plotting points, Default: 100}
}
\value{
None.
}
\description{
Check Probability Conservation in Standardizing the Normal Distribution
}
\examples{
norm.trans(2, 2, -1, 4)
}
