% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normal_distribution.R
\name{normal_distribution}
\alias{normal_distribution}
\title{Normalna distribucija (interaktivni unos)}
\usage{
normal_distribution()
}
\value{
Vraća \code{plotly} graf normalne distribucije sa osenčenim područjem
do unesene vrijednosti x.
}
\description{
Funkcija prikazuje graf normalne distribucije uz zadane parametre
aritmetičke sredine (μ), standardne devijacije (σ) i vrijednosti x.
Parametri se unose interaktivno putem konzole.
}
\details{
Površina do vrijednosti x je osenčena, a na grafu se prikazuju
vertikalne linije za μ i x. U konzoli se ispisuju vjerojatnosti
P(X ≤ x) i P(X ≥ x).

Funkcija koristi paket \code{plotly}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
}
\examples{
\dontrun{
normal_distribution()
}

}
