% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normality_tests.R
\name{normality_tests}
\alias{normality_tests}
\title{Testovi normalnosti za numeričke varijable}
\usage{
normality_tests(df, interpret = TRUE)
}
\arguments{
\item{df}{\code{data.frame} koji sadrži numeričke varijable za testiranje.}

\item{interpret}{Logička vrijednost. Ako je TRUE (default), funkcija dodaje kolonu s interpretacijom rezultata (normalno / nije normalno).}
}
\value{
Lista sa sljedećim komponentama:
\itemize{
\item \strong{table} - tablica rezultata testa za svaku odabranu varijablu
\item \strong{test} - naziv odabranog testa
}
}
\description{
Funkcija \code{normality_tests} omogućuje odabir jedne ili više numeričkih varijabli iz
\code{data.frame}-a i provođenje odabranog testa normalnosti: Kolmogorov-Smirnov,
Shapiro-Wilk, Lilliefors ili Anderson-Darling. Također može dodati interpretaciju rezultata.
}
\details{
Funkcija koristi paket \code{nortest}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
Varijable se odabiru putem interaktivnog izbornika. Shapiro-Wilk test nije primjenjiv za više od 5000 entiteta i tada vraća NA.
}
\examples{
\dontrun{
# Primjer korištenja s data.frame-om df
normality_tests(df)
}

}
