% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{normtest.plot}
\alias{normtest.plot}
\title{Plot the PDF of Test Statistic under the Normal Distribution}
\usage{
normtest.plot(md, mu0 = 0, se = 1, prng = c(-4, 4), side = "two",
  xlab = "Sample Mean", pvout = TRUE)
}
\arguments{
\item{md}{Difference of sample means or test statistic}

\item{mu0}{Difference of population means under the null hypothesis, Default: 0}

\item{se}{Standard error of the difference of sample means, Default: 1}

\item{prng}{Range of x-axis, Default: c(-4, 4)}

\item{side}{Type of the alternative hypothesis, Default: 'two'}

\item{xlab}{Label of x-axis, Default: 'Sample Mean'}

\item{pvout}{Print p-value? Default: TRUE}
}
\value{
None.
}
\description{
Plot the PDF of Test Statistic under the Normal Distribution
}
\examples{
normtest.plot(11.5-10.4, se=sqrt(2.2/20+2.4/25))
normtest.plot(1.93, xlab="Test Statistic")
}
