% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{pmean.ci}
\alias{pmean.ci}
\title{Confidence Interval for a Mean (known variance)}
\usage{
pmean.ci(xb, sig, n, alp = 0.05, dig = 4)
}
\arguments{
\item{xb}{Sample mean, or sample data}

\item{sig}{Population standard deviation}

\item{n}{Sample size (unnecessary when sample data are given)}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for a Population Mean (known variance)
}
\examples{
pmean.ci(xb=199.5, sig=5, n=50)
pmean.ci(xb=199.5, sig=5, n=50, dig=3)
pmean.ci(rnorm(100, 20, 3), sig=3)
}
