% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{pmean.ci2}
\alias{pmean.ci2}
\title{Confidence Interval for the Mean (unknown variance)}
\usage{
pmean.ci2(xb, sig, n, alp = 0.05, dig = 4)
}
\arguments{
\item{xb}{Sample mean, or sample data}

\item{sig}{Sample standard deviation (unnecessary when sample data are given)}

\item{n}{Sample size (unnecessary when sample data are given)}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for the Population Mean (unknown variance)
}
\examples{
pmean.ci2(xb=199.5, sig=5, n=16, alp=0.05, dig=3)
pmean.ci2(rnorm(20, 199.5, 5))
}
