% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{pprt}
\alias{pprt}
\title{Probability of an Event}
\usage{
pprt(x, n, prt = TRUE)
}
\arguments{
\item{x}{Event in data frame.}

\item{n}{Size of the sample space.}

\item{prt}{Print output? Default: TRUE}
}
\value{
The probability of an event.
}
\description{
Calculating the Probability of an Event
}
\examples{
S = rolldie2(2)
B = subset(S, (X1+X2) >=8)
pprt(B, nrow(S))
}
