% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{prob.ci}
\alias{prob.ci}
\title{Confidence Interval for a Population Proportion}
\usage{
prob.ci(n, x, alp = 0.05, dig = 4)
}
\arguments{
\item{n}{Sample size}

\item{x}{Number of successes in a sample}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for a Population Proportion
}
\examples{
prob.ci(n=200, x=15)
}
