% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{prob2.ci}
\alias{prob2.ci}
\title{Confidence Interval for the Difference of Proportions}
\usage{
prob2.ci(n1, x1, n2, x2, alp = 0.05, dig = 4)
}
\arguments{
\item{n1}{Sample size of population1}

\item{x1}{Number of successes in samples from population1}

\item{n2}{Sample size of population2}

\item{x2}{Number of successes in samples from population2}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for the Difference of Population Proportions
}
\examples{
prob2.ci(n1=160, x1=12, n2=200, x2=13, alp=0.1)
prob2.ci(n1=160, x1=12, n2=200, x2=13, alp=0.05)
}
