% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{prob2test.plot}
\alias{prob2test.plot}
\title{Test for the Difference of Proportions (Large Sample)}
\usage{
prob2test.plot(n1, x1, n2, x2, prng, side = "two", mt, dig = 4)
}
\arguments{
\item{n1}{Sample size of population1}

\item{x1}{Number of successes in samples from population1}

\item{n2}{Sample size of population2}

\item{x2}{Number of successes in samples from population2}

\item{prng}{Range of x-axis, Default: [d0-4se, d0+4se]}

\item{side}{Type of the alternative hypothesis, Default: 'two'}

\item{mt}{Graph title}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Test for the Difference of Population Proportions (Large Sample)
}
\examples{
prob2test.plot(n1=150, x1=12, n2=250, x2=10, side="up", dig=4)
prob2test.plot(n1=150, x1=12, n2=250, x2=10, side="two", dig=4)
}
