% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{ranksum.dist}
\alias{ranksum.dist}
\title{Distribution of Wilcoxon Rank Sum}
\usage{
ranksum.dist(n1, n2 = 3:10, tab = TRUE, plot = FALSE, dig = 4)
}
\arguments{
\item{n1}{Number of data in group 1 (default=1:min(n2))}

\item{n2}{Number of data in group 2, Default: 3:10}

\item{tab}{Print critical value table? Default: TRUE}

\item{plot}{Plot rank sum distribution? Default: FALSE}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Distribution of Wilcoxon Rank Sum Test Statistic
}
\examples{
ranksum.dist(n2=3:5)
ranksum.dist(n1=c(2,4,6,8), n2=10, tab=FALSE, plot=TRUE)
}
