% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{ranksum.plot}
\alias{ranksum.plot}
\title{Wilcoxon Rank Sum Test}
\usage{
ranksum.plot(x, y, side = "two", xlab = "Rank Sum statistic",
  dig = 4)
}
\arguments{
\item{x}{Data vector in group 1}

\item{y}{Data vector in group 2}

\item{side}{Type of alternative hypothesis, Default: 'two'}

\item{xlab}{Label of x-axis, Default: 'Rank Sum statistic'}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Wilcoxon Rank Sum Test with a Plot
}
\examples{
x = c(1,5,7,8,8,8,9)
y = c(2,8,8,8,8,9,9,9,9,10,10,10,10)
ranksum.plot(x, y)
}
