% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch4-fn.R
\name{rolldie.sum}
\alias{rolldie.sum}
\title{PDF of the Sum of n Dice}
\usage{
rolldie.sum(n, cex = 1)
}
\arguments{
\item{n}{Number of dice.}

\item{cex}{Default: 1}
}
\value{
None.
}
\description{
Probability Distribution of the Sum of n Dice
}
\examples{
rolldie.sum(4)
}
