% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{rolldie2}
\alias{rolldie2}
\title{Sample Space of Rolling Dice}
\usage{
rolldie2(times, nsides = 6)
}
\arguments{
\item{times}{Number of dice}

\item{nsides}{Number of sides of a die, Default: 6}
}
\value{
Sample space in data frame
}
\description{
Create Sample Space of Rolling Dice
}
\examples{
rolldie2(2)
}
