% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runs.dist}
\alias{runs.dist}
\title{Distribution of Runs}
\usage{
runs.dist(n1 = 2:20, n2 = 2:20, alp = 0.05, tab = TRUE,
  side = "two", plot = FALSE)
}
\arguments{
\item{n1}{Number of data in group 1, Default: 2:20}

\item{n2}{Number of data in group 2, Default: 2:20}

\item{alp}{Level of significance, Default: 0.05}

\item{tab}{Print critical value table? Default: TRUE}

\item{side}{Type of alternative hypothesis, Default: 'two'}

\item{plot}{Plot run distribution? Default: FALSE}
}
\value{
None.
}
\description{
Distribution of Runs Test Statistic
}
\examples{
require(randomizeBE)
runs.dist(n1=2:10, n2=2:10)
runs.dist(n1=c(5,20), n2=c(5,20), tab=FALSE, plot=TRUE)
}
