% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{runstest.plot}
\alias{runstest.plot}
\title{Runs Test}
\usage{
runstest.plot(x, n1, n2, alp = 0.05, side = "two", dig = 4,
  plot = TRUE)
}
\arguments{
\item{x}{Data vector (or number of runs)}

\item{n1}{Number of data in group 1 (required if raw data are not given)}

\item{n2}{Number of data in group 2 (required if raw data are not given)}

\item{alp}{Level of significance, Default: 0.05}

\item{side}{Type of alternative hypothesis, Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{plot}{Plot runs test results? Default: TRUE}
}
\value{
None.
}
\description{
Runs Test with a Plot
}
\examples{
require(randomizeBE)
x = c(1,1,0,0,1,0,rep(1,7), rep(0,7))
runstest.plot(x)

x = rep(0, 50)
x[c(1:2, 8:10, 17:18, 21, 26:27, 29:31, 36:37, 41:44, 49)] = 1
runstest.plot(x)
}
