% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{scat.lm}
\alias{scat.lm}
\title{Scatter Plot with a Regression Line}
\usage{
scat.lm(x, y, mt, xl, yl, w = c(7, 5), ...)
}
\arguments{
\item{x}{Vector of x variable.}

\item{y}{Vector of y variable.}

\item{mt}{Title of the plot.}

\item{xl}{Label of x-axis.}

\item{yl}{Label of y-axis.}

\item{w}{Size of graphic window. Default: c(7, 5)}

\item{...}{Other graphic parameters.}
}
\value{
None.
}
\description{
To create a scatter plot with a linear regression line.
}
\examples{
scat.lm(mtcars$wt, mtcars$mpg, mt="Weight vs. MPG")
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
