% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signrank.dist}
\alias{signrank.dist}
\title{Distribution of Wilcoxon Signed Rank}
\usage{
signrank.dist(nv = 5:50, av, tab = TRUE, plot = FALSE, dig = 4)
}
\arguments{
\item{nv}{Number of signed data, Default: 5:50}

\item{av}{Probability vector (default=c(0.005, 0.01, 0.025, 0.05, 0.95, 0.975, 0.99, 0.995))}

\item{tab}{Print quantile table? Default: TRUE}

\item{plot}{Plot test statistic distribution? Default: FALSE}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Distribution of Wilcoxon Signed Rank Test Statistic
}
\examples{
signrank.dist(nv=5:15)
signrank.dist(nv=c(5,7,10,20), tab=FALSE, plot=TRUE)
}
