% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signrank.plot}
\alias{signrank.plot}
\title{Wilcoxon Signed Rank Test}
\usage{
signrank.plot(x, y, mu0 = 0, side = "two", xlab = "Signed Rank Sum",
  dig = 4)
}
\arguments{
\item{x}{Data vector in group 1}

\item{y}{Data vector in group 2}

\item{mu0}{Mean difference under the null hypothesis, Default: 0}

\item{side}{Type of alternative hypothesis, Default: 'two'}

\item{xlab}{Label of x-axis, Default: 'Signed Rank Sum'}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Wilcoxon Signed Rank Test with a Plot
}
\examples{
x = c(38, 26, 34, 5, 68, 30, 35, 19, 33, 69)
y = c(28, 21, 31, 11, 59, 28, 28, 23, 32, 38)
signrank.plot(x=x, y=y, side="up")
}
