% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch15-fn.R
\name{signtest.plot}
\alias{signtest.plot}
\title{Sign Test (Binomial Test)}
\usage{
signtest.plot(x, mu0 = 0, side = "two", dig = 4)
}
\arguments{
\item{x}{Data vector}

\item{mu0}{Mean value under the null hypothesis, Default: 0}

\item{side}{Type of alternative hypothesis, Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Sign Test (Binomial Test) with a Plot
}
\examples{
(x = c(1,2,5,7, rep(8,7), rep(9,5), rep(10,4)))
signtest.plot(x=x, mu0=7, side="up")
}
