% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpl_reg.R
\name{simpl_reg}
\alias{simpl_reg}
\title{Jednostavna linearna regresijska analiza}
\usage{
simpl_reg(df)
}
\arguments{
\item{df}{Data frame koji sadrži najmanje dvije numeričke varijable.
Funkcija omogućuje korisniku da odabere nezavisnu i zavisnu varijablu.}
}
\value{
Lista koja sadrži:
\item{b0}{Intercept (b0) regresije.}
\item{b1}{Koeficijent nagiba (b1).}
\item{r}{Korelacijski koeficijent .}
\item{see}{Standardna greška prognoze.}
\item{F}{F-vrijednost.}
\item{df1}{Stupnjevi slobode (df1).}
\item{df2}{Stupnjevi slobode (df2).}
\item{p}{P-vrijednost F-testa.}
\item{residuals}{Vektor residuala vrijednosti}
\item{fitted}{Vektor prognoziranih vrijednosti.}
\item{graph1}{Interaktivni scatter plot sa regresijskom linijom (plotly).}
\item{graph2}{Interaktivni grafikon residuala (plotly).}

#' @details
Ako paket \code{plotly} nije instaliran, funkcija ga automatski instalira
i učitava. Odabir varijabli je interaktivan putem \code{select.list}.
}
\description{
Funkcija provodi jednostavnu linearnu regresiju između dvije numeričke varijable
iz zadanog data frame-a i vraća parametre regresijskog modela i grafikone koristeći paket \code{plotly}.
}
\examples{
\dontrun{
result <- simpl_reg(mtcars)
result$graph1  # prikaz scatter plot s regresijskom linijom
result$graph2  # prikaz residualnog grafika
}

}
