% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.cdf}
\alias{snorm.cdf}
\title{Plot Standard Normal Cumulative Probability}
\usage{
snorm.cdf(zp, lo = -4, up = 4, mt, dig = 4)
}
\arguments{
\item{zp}{Vector of z-axis values (default=-2:2)}

\item{lo}{Lower limit of z-axis, Default: -4}

\item{up}{Upper limit of z-axis, Default: 4}

\item{mt}{Graph title, Default: 'Cumulative Probabilities of the Standard Normal Distribution'}

\item{dig}{Number of digits below the decimal point (default=4), Default: 4}
}
\value{
None.
}
\description{
Plot Standard Normal Cumulative Probability P(Z<z)
}
\examples{
zp = seq(-2, 2, by=0.5)
snorm.cdf(zp)
}
