% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.prob}
\alias{snorm.prob}
\title{Central Probability of the Standard Normal Distribution}
\usage{
snorm.prob(zp, lo = -4, up = 4, mt, dig = 4)
}
\arguments{
\item{zp}{Vector of z-axis values (default=1:4)}

\item{lo}{Lower limit of z-axis, Default: -4}

\item{up}{Upper limit of z-axis, Default: 4}

\item{mt}{Graph title (default="Central Probability of the Standard Normal Distribution")}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Central Probability of the Standard Normal Distribution
}
\examples{
zp = 1:3
snorm.prob(zp)
}
