% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch8-fn.R
\name{snorm.quant}
\alias{snorm.quant}
\title{Quantile Plot of the Standard Normal Distribution}
\usage{
snorm.quant(pv, pv2, mt, dig = 4)
}
\arguments{
\item{pv}{Vector of probability values}

\item{pv2}{Vector of specific probability values}

\item{mt}{Graph title (default="Quantiles of the Standard Normal Distribution")}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Quantile Plot of the Standard Normal Distribution
}
\examples{
pv = c(0.005, 0.01, 0.025, 0.05, 1:9/10, 0.95, 0.975, 0.99, 0.995)
snorm.quant(pv, pv)
}
