% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{spread.est}
\alias{spread.est}
\title{Spread Estimation}
\usage{
spread.est(x, detail = FALSE)
}
\arguments{
\item{x}{Vector of input data.}

\item{detail}{Print detailed output? Default: FALSE}
}
\value{
None
}
\description{
To calculate spread estimates.
}
\examples{
spread.est(mtcars$mpg, detail=T)
}
\keyword{Ch2.}
\keyword{Descriptive}
\keyword{Statistics}
