% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{strat.hist}
\alias{strat.hist}
\title{Stratified Histograms}
\usage{
strat.hist(ng = 3, n = 200, m = c(6, 10, 14), s = 1, sp = c(4,
  16), vc, prob = FALSE)
}
\arguments{
\item{ng}{Number of classes. Defaults to 3.}

\item{n}{Number of samples. Defaults to 200.}

\item{m}{Vector of means. Defaults to c(6,10,14)}

\item{s}{Standard deviations. Default: 1}

\item{sp}{Specification limits. Default: c(4, 16)}

\item{vc}{Colors of histograms.}

\item{prob}{Density scale? Default: FALSE}
}
\value{
None
}
\description{
To create Stratified Histograms from a mixed data set
}
\examples{
strat.hist()
strat.hist(ng=4, m=c(6,10,14,18), sp=c(4,20))
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics}
