% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_dep.R
\name{t_test_dep}
\alias{t_test_dep}
\title{T-test za zavisne uzorke za više varijabli}
\usage{
t_test_dep(df)
}
\arguments{
\item{df}{Data frame koji sadrži varijable za analizu.}
}
\value{
Data frame sa sljedećim kolonama:
\itemize{
\item VARS1 - naziv varijable prvog mjerenja
\item VARS2 - naziv varijable drugog mjerenja
\item MEAN1 - aritmetička sredina prvog mjerenja
\item MEAN2 - aritmetička sredina drugog mjerenja
\item \code{MEAN-DIF} - razlika aritmetičkh sredina
\item SED - standardna pogreška razlike
\item T - t-statistika
\item DF - stupnjevi slobode
\item P - p-vrijednost
\item \verb{Cohen's D} - Cohenov d
}
}
\description{
Ova funkcija omogućuje izvođenje t-testova zavisne uzorke za odabrane parove varijabli u data frame-u.
Također izračunava aritmetičku sredinu, standardnu pogrešku razlike, t-vrijednost, stupnjeve slobode,
p-vrijednost i Cohenov d.
}
\details{
Funkcija koristi interaktivni izbor varijabli prvog i drugog mjerenja.
Broj varijabli u prvom i drugom mjerenju mora biti jednak.
Funkcija koristi paket \code{lsr}. Ako paket nije instaliran, funkcija će ga automatski instalirati i učitati.
}
\examples{
\dontrun{
rezultati <- t_test_dep(df)
}

}
