% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_ind.R
\name{t_test_ind}
\alias{t_test_ind}
\title{T-test za nezavisne uzorke za više varijabli}
\usage{
t_test_ind(df)
}
\arguments{
\item{df}{Data frame koji sadrži podatke.}
}
\value{
Data frame s rezultatima t-testa:
\itemize{
\item aritmetičke sredine po grupama,
\item standardne pogreške razlika (SED),
\item t-vrijednosti i stupnjeve slobode (DF),
\item p-vrijednost,
\item 95\% interval pouzdanosti,
\item Levene F i p-vrijednosti,
\item Cohenov d,
\item vrstu t-testa (Student/Welch).
}
}
\description{
Ova funkcija izvodi t-test za nezavisne uzorke (Student, Welch ili automatski prema Levene testu)
za više kvantitativnih varijabli i jednu kvalitativnu varijablu s dvije kategorije.
}
\details{
Funkcija omogućava interaktivni odabir:
\itemize{
\item kvalitativne (grupne) varijable,
\item jedne ili više kvantitativnih varijabli,
\item tip t-testa: "Student t-test", "Welch t-test" ili "Auto" (odabire se prema Levene testu za jednakost varijanci).
}
Funkcija koristi pakete \code{lsr} i \code{car}. Ako paketi nusu instalirani, funkcija će ih automatski instalirati i učitati.
}
\examples{
\dontrun{
library(readxl)
df <- read_excel(file.choose())
rezultati <- t_test_ind(df)
print(rezultati)
}

}
