% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{tosscoin2}
\alias{tosscoin2}
\title{Sample Space of Tossing Coins}
\usage{
tosscoin2(times)
}
\arguments{
\item{times}{Number of coins}
}
\value{
Sample space in data frame
}
\description{
Create Sample Space of Tossing Coins
}
\examples{
tosscoin2(4)
}
