% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trans_value.R
\name{trans_value}
\alias{trans_value}
\title{Transformacije podataka}
\usage{
trans_value(df)
}
\arguments{
\item{df}{DataFrame koji sadrži numeričke varijable.}
}
\value{
Lista s dva elementa:
\item{table}{DataFrame s transformiranim vrijednostima}
\item{method}{Odabrana metoda transformacije}
}
\description{
Funkcija transformaciju numeričkih varijabli u Z-value, T-value ili L-value.
Omogućuje odabir normalno i obrnuto skaliranih varijabli.
}
\details{
Normalne varijable se standardiziraju (Z-score), a obrnuto skalirane varijable
se množe s -1. T-value transformacija: T = Z * 10 + 50. L-value transformacija: L = Z * 0.83 + 3
}
\examples{
# result <- trans_value(df)

}
