% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{ttest.plot}
\alias{ttest.plot}
\title{Plot the PDF of Test Statistic with the T-distribution}
\usage{
ttest.plot(md, deg, prng = c(-4, 4), side = "two", dig = 4, mt,
  pvout = TRUE)
}
\arguments{
\item{md}{T-test statistic for the difference of population means}

\item{deg}{Degree of freedom}

\item{prng}{Range of x-axis, Default: c(-4, 4)}

\item{side}{Type of the alternative hypothesis, Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}

\item{mt}{Plot title}

\item{pvout}{Print p-value? Default: TRUE}
}
\value{
None.
}
\description{
Plot the PDF of Test Statistic with the T-distribution
}
\examples{
ttest.plot(1.96, deg=24)
}
