% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch2-fn.R
\name{unstable.hist}
\alias{unstable.hist}
\title{Unstable Histograms}
\usage{
unstable.hist(N = 200, m1 = 10, s1 = 1, m2 = 6, s2 = 0.5,
  a = 8, b = 9, c = 9, vc = rep("cyan", 4))
}
\arguments{
\item{N}{Number of random variates. Defaults to 200.}

\item{m1, m2}{Means of two normal distributions.}

\item{s1, s2}{Standard deviations of two normal distributions.}

\item{mp}{Logical variable for making histogram. Defaults to FALSE.}

\item{...}{Other graphic parameters}
}
\value{
None
}
\description{
To create 4 types of unstable histograms
}
\examples{
unstable.hist()
unstable.hist(m1=15, s1=2, m2=8, s2=1, a=12, b=13, c=10)
}
\keyword{Ch2}
\keyword{Descriptive}
\keyword{Statistics}
