% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch3-fn.R
\name{urnsample2}
\alias{urnsample2}
\title{Sample Space of Urn Sampling}
\usage{
urnsample2(x, size, replace = FALSE, ordered = FALSE,
  probspace = FALSE, ...)
}
\arguments{
\item{x}{Vector of objects in the urn}

\item{size}{Number of samples from the urn}

\item{replace}{Sampling with replacement? Default: FALSE}

\item{ordered}{Consider the order of samples? Default: FALSE}

\item{probspace}{Create probablity space? Default: FALSE}

\item{...}{Other parameters}
}
\value{
Sample space in data frame
}
\description{
Create Sample Space of Urn Sampling
}
\examples{
urnsample2(letters[1:5], 3)
}
