% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{var1.ci}
\alias{var1.ci}
\title{Confidence Interval for a Variance}
\usage{
var1.ci(x, alp = 0.05, dig = 4)
}
\arguments{
\item{x}{Data vector}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for a Population Variance
}
\examples{
x = c(20.0, 21.5, 20.9, 19.8, 22.5, 20.3, 23.6, 18.0, 23.3, 17.8)
var1.ci(x, dig=3)
var1.ci(rnorm(36))
}
