% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch10-fn.R
\name{var1.test}
\alias{var1.test}
\title{Hypothesis Test for a Variance}
\usage{
var1.test(x, n, var0, alp = 0.05, side = "two", dig = 4)
}
\arguments{
\item{x}{Data vector (or sample variance)}

\item{n}{Sample size (necessary when the sample variance is given)}

\item{var0}{Population variance value under the null hypothesis}

\item{alp}{Level of significance, Default: 0.05}

\item{side}{Type of the alternative hypothesis ("up", "low", "two"), Default: 'two'}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Hypothesis Test for a Population Variance
}
\examples{
var1.test(x=1.24, n=25, var0=0.8, side="up")

x = c(20.0, 21.5, 20.9, 19.8, 22.5, 20.3, 23.6, 18.0, 23.3, 17.8)
var1.test(x=x, var0=2, side="two")
}
