% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch11-fn.R
\name{var2.ci}
\alias{var2.ci}
\title{CI for the Ratio of Two Population Variances}
\usage{
var2.ci(s1, s2, n1, n2, alp = 0.05, dig = 4)
}
\arguments{
\item{s1}{Sample standard deviation of population1}

\item{s2}{Sample standard deviation of population2}

\item{n1}{Sample size of population1}

\item{n2}{Sample size of population2}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Confidence Interval for the Ratio of Two Population Variances
}
\examples{
var2.ci(s1=3.5, s2=5.8, n1=25, n2=34)

x = rnorm(20, 199, 4)
y = rnorm(25, 200, 2)
var2.ci(sd(x), sd(y), length(x), length(y))
}
