% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ch14-fn.R
\name{corr.reg1}
\alias{corr.reg1}
\title{Correlation & Simple Regression Analysis}
\usage{
corr.reg1(x, y, r0 = 0, xl, yl, mt, step = 1:4, x0, xrng, by,
  alp = 0.05, dig = 4)
}
\arguments{
\item{x}{Vector of independent variable (explanatory variable) data}

\item{y}{Vector of dependent variable (response variable) data}

\item{r0}{, Default: 0}

\item{xl}{Label of x-axis (default=x variable name)}

\item{yl}{Label of y-axis (default=y variable name)}

\item{mt}{Plot title}

\item{step}{Steps of simple regression analysis, Default: 1:9}

\item{x0}{Specipic poit value for confidence and prediction intervals}

\item{xrng}{Range of x-axis}

\item{by}{Plotting interavl of confidence and prediction bands}

\item{alp}{Level of significance, Default: 0.05}

\item{dig}{Number of digits below the decimal point, Default: 4}
}
\value{
None.
}
\description{
Correlation Analysis & Simple Linear Regression Analysis
}
\examples{
corr.reg1(iris[[3]], iris[[4]])
corr.reg1(iris[[3]], iris[[4]], step=6)
corr.reg1(iris[[3]], iris[[4]], x0=3, step=11)
}
